+++ itolib/Behaviours/Enemies/EnemySpawner.cs
@@ -1,5 +1,6 @@
 using itolib.Extensions;
 using itolib.Interfaces;
+using itolib.Structs;
 using itolib.Util;
 using System;
 using Unity.Netcode;
@@ -8,101 +9,95 @@ using UnityEngine;
 namespace itolib.Behaviours.Enemies
 {
     /// <summary>
-    ///     TODO.
+    ///     Represents a single entry with weights to be used for weighted enemy selection.
     /// </summary>
-    /// <param name="enemyWithRarity"></param>
+    /// <param name="enemyWithRarity">Enemy to copy name and weights from.</param>
     [Serializable]
     public struct EnemyWeightEntry(SpawnableEnemyWithRarity enemyWithRarity) : IWeightedEntry
     {
         /// <summary>
-        ///     TODO.
+        ///     Enemy name corresponding to this specific entry.
         /// </summary>
         [Header("Enemy Weight Entry")]
-        [Tooltip("")]
+        [Tooltip("Enemy name corresponding to this specific entry.")]
         public string enemyName = (enemyWithRarity.enemyType != null) ? enemyWithRarity.enemyType.enemyName : string.Empty;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        [field: Tooltip("")]
+        /// <inheritdoc/>
+        [field: Tooltip("Weight value for this specific entry.")]
         [field: Min(0)]
         [field: SerializeField] public int Weight { get; set; } = enemyWithRarity.rarity;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        [field: Tooltip("")]
+        /// <inheritdoc/>
+        [field: Tooltip("Weight modifiers to apply whenever this specific entry is used.")]
+        [field: SerializeField] public WeightedModifier[]? WeightedModifiers { get; set; }
+
+        /// <inheritdoc/>
+        [field: Tooltip("Whether this specific entry can be used more than once or not.")]
         [field: SerializeField] public bool SingleUse { get; set; } = false;
 
         /// <summary>
-        ///     TODO.
+        ///     Enemy corresponding to this specific entry.
         /// </summary>
+        /// <remarks>Deprecated. Should be ignored.</remarks>
         [Space(5.0f)]
         [Header("== DEPRECATED ==")]
         [Tooltip("(Deprecated) Replace with the desired enemy's 'enemyName' field.")]
+        [Obsolete("Replace with the desired enemy's 'enemyName' field.")]
         public EnemyType? enemyToSpawn = enemyWithRarity.enemyType;
     }
 
     /// <summary>
-    ///     TODO.
+    ///     Represents an enemy spawner with weighted selection capabilities.
     /// </summary>
     public class EnemySpawner : EnemySpawnerBase<EnemyAI>, IWeightedScript<EnemyWeightEntry>
     {
-        /// <summary>
-        ///     Cached instance of <c>EnemySpawner</c> as an <c>IWeightedScript</c>, to avoid having to cast.
-        /// </summary>
+        /// <inheritdoc/>
         public IWeightedScript<EnemyWeightEntry> WeightedSelf { get; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        public int[]? CumulativeWeights { get; set; }
+        /// <inheritdoc/>
+        public int[]? CurrentWeights { get; set; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
+        /// <inheritdoc/>
         public int TotalWeight { get; set; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
+        /// <inheritdoc/>
         public bool InitializedWeights { get; set; }
 
         /// <summary>
-        ///     TODO.
+        ///     Name of the enemy to be spawned.
         /// </summary>
+        /// <remarks><b>NOTE:</b> Should be left empty if intending to use weighted selection.</remarks>
         [Space(5.0f)]
         [Header("Enemy Spawner")]
-        [Tooltip("")]
+        [Tooltip("Name of the enemy to be spawned. NOTE: Should be left empty if intending to use weighted selection.")]
         [SerializeField] private string enemyToSpawn = string.Empty;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        [field: Tooltip("")]
+        /// <inheritdoc/>
+        [field: Tooltip("List of weighted entries of type EnemyWeightEntry.")]
         [field: SerializeField] public EnemyWeightEntry[]? WeightedEntries { get; set; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc/>
         public override NetworkObject? GetPrefabToSpawn()
         {
+            // Spawn enemy specified in the 'enemyToSpawn' field, if one is set.
             if (!enemyToSpawn.IsNullOrEmpty() && SearchContent.TryFindEnemy(out EnemyType enemy, enemyToSpawn)
                 && TryGetNetworkObject(out NetworkObject enemyNetworkObject, enemy))
             {
                 return enemyNetworkObject;
             }
 
-            if (WeightedEntries?.Length > 0 && WeightedSelf.TryObtainRandomEntry(out EnemyWeightEntry entry, isSeededRandom
+            // Spawn enemy using weighted selection.
+            if (WeightedEntries?.Length > 0 && WeightedSelf.TryObtainRandomEntry(out EnemyWeightEntry entry, out int _, isSeededRandom
                 ? SeededSelf.GetSeededRandom() : null))
             {
                 if (SearchContent.TryFindEnemy(out enemy, entry.enemyName) && TryGetNetworkObject(out enemyNetworkObject, enemy))
                 {
                     return enemyNetworkObject;
                 }
+#pragma warning disable CS0618 // Type or member is obsolete.
                 else if (entry.enemyToSpawn != null && (TryGetNetworkObject(out enemyNetworkObject, entry.enemyToSpawn) || (SearchContent.TryFindEnemy(out enemy,
                     entry.enemyToSpawn.name, checkObjectName: true) && TryGetNetworkObject(out enemyNetworkObject, enemy))))
+#pragma warning restore CS0618 // Type or member is obsolete.
                 {
                     return enemyNetworkObject;
                 }
@@ -120,7 +115,7 @@ namespace itolib.Behaviours.Enemies
         }
 
         /// <summary>
-        ///     Initialize weights for every <c>EnemyWeightEntry</c>.
+        ///     Initialize weights for every defined <c>EnemyWeightEntry</c>.
         /// </summary>
         protected override void Awake()
         {
@@ -133,9 +128,9 @@ namespace itolib.Behaviours.Enemies
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Add a single weighted entry of type <c>EnemyWeightEntry</c>.
         /// </summary>
-        /// <param name="entry"></param>
+        /// <param name="entry">Entry of type <c>EnemyWeightEntry</c> to add.</param>
         public void AddWeightEntry(EnemyWeightEntry entry)
         {
             if (!WeightedSelf.InitializedWeights)
@@ -147,9 +142,9 @@ namespace itolib.Behaviours.Enemies
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Add multiple weighted entries of type <c>EnemyWeightEntry></c>.
         /// </summary>
-        /// <param name="entries"></param>
+        /// <param name="entries">Entries of type <c>EnemyWeightEntry</c> to add.</param>
         public void AddWeightEntries(EnemyWeightEntry[] entries)
         {
             if (!WeightedSelf.InitializedWeights)
@@ -161,9 +156,10 @@ namespace itolib.Behaviours.Enemies
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Remove weights for the weighted entry of type <c>EnemyWeightEntry</c> at the specified index.
         /// </summary>
-        /// <param name="index"></param>
+        /// <remarks>Sets weights to <c>0</c> instead of actually removing them.</remarks>
+        /// <param name="index">Index of the entry of type <c>EnemyWeightEntry</c> to remove.</param>
         public void RemoveWeightEntry(int index)
         {
             if (!WeightedSelf.InitializedWeights)
@@ -175,9 +171,10 @@ namespace itolib.Behaviours.Enemies
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Switch target enemy to be spawned.
         /// </summary>
-        /// <param name="enemyName"></param>
+        /// <remarks><b>NOTE:</b> Should not be used if intending to use weighted selection.</remarks>
+        /// <param name="enemyName">Name of the enemy to be spawned.</param>
         public void SwitchEnemyToSpawn(string enemyName)
         {
             enemyToSpawn = enemyName;