+++ itolib/Behaviours/Enemies/EnemySpawnerBase.cs
@@ -6,24 +6,22 @@ using UnityEngine;
 namespace itolib.Behaviours.Enemies
 {
     /// <summary>
-    ///     TODO.
+    ///     Represents an abstract enemy spawner.
     /// </summary>
-    /// <typeparam name="T"></typeparam>
+    /// <typeparam name="T">Enemy of type <c>EnemyAI</c> to be spawned.</typeparam>
     public abstract class EnemySpawnerBase<T> : NetworkedSpawner<T> where T : EnemyAI
     {
         /// <summary>
-        ///     TODO.
+        ///     Whether spawned enemies should increase or decrease enemy power levels for the current round or not.
         /// </summary>
+        /// <remarks><b>NOTE:</b> Affected power levels may not correspond to where enemies are actually spawned.</remarks>
         [Space(5.0f)]
         [Header("Enemy Spawner Base")]
-        [Tooltip("")]
+        [Tooltip("Whether spawned enemies should increase or decrease enemy power levels for the current round or not. NOTE: Affected power levels "
+            + "may not correspond to where enemies are actually spawned.")]
         [SerializeField] private bool influencePowerLevels = true;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        /// <param name="spawnedPrefab"></param>
-        /// <param name="spawnLocation"></param>
+        /// <inheritdoc/>
         protected override void SpawnPerformed(T? spawnedPrefab, TransformInfo spawnLocation)
         {
             if (spawnedPrefab == null || !spawnedPrefab.IsSpawned)
@@ -33,6 +31,7 @@ namespace itolib.Behaviours.Enemies
 
             if (RoundManager.Instance != null)
             {
+                // Apply enemy power level changes.
                 if (influencePowerLevels)
                 {
                     EnemyType? enemyType = spawnedPrefab.enemyType;
@@ -55,6 +54,7 @@ namespace itolib.Behaviours.Enemies
                 }
                 else
                 {
+                    // Set power level as removed to not decrease it when the enemy dies.
                     spawnedPrefab.removedPowerLevel = true;
                 }
 
@@ -65,11 +65,11 @@ namespace itolib.Behaviours.Enemies
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Attempt to obtain the <c>NetworkObject</c> to spawn from a given <c>EnemyType</c>.
         /// </summary>
-        /// <param name="enemyNetworkObject"></param>
-        /// <param name="enemy"></param>
-        /// <returns></returns>
+        /// <param name="enemyNetworkObject"><c>NetworkObject</c> of the enemy to spawn, as an out parameter.</param>
+        /// <param name="enemy">Enemy to obtain the <c>NetworkObject</c> off of.</param>
+        /// <returns>Whether a <c>NetworkObject</c> was successfully obtained or not.</returns>
         protected static bool TryGetNetworkObject(out NetworkObject enemyNetworkObject, EnemyType? enemy)
         {
             enemyNetworkObject = null!;