+++ itolib/Behaviours/Props/ScrapSpawner.cs
@@ -2,7 +2,6 @@ using itolib.Behaviours.Networking;
 using itolib.Interfaces;
 using itolib.Structs;
 using itolib.Util;
-using LethalLevelLoader;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -13,182 +12,178 @@ using UnityEngine.Serialization;
 namespace itolib.Behaviours.Props
 {
     /// <summary>
-    ///     TODO.
+    ///     Represents a single entry with weights to be used for weighted scrap item selection.
     /// </summary>
-    /// <param name="itemWithRarity"></param>
+    /// <param name="itemWithRarity">Scrap item to copy name and weights from.</param>
     [Serializable]
     public struct ScrapWeightEntry(SpawnableItemWithRarity itemWithRarity) : IWeightedEntry
     {
         /// <summary>
-        ///     TODO.
+        ///     Scrap item name corresponding to this specific entry.
         /// </summary>
         [Header("Scrap Weight Entry")]
-        [Tooltip("")]
+        [Tooltip("Scrap item name corresponding to this specific entry.")]
         public string itemName = (itemWithRarity.spawnableItem != null) ? itemWithRarity.spawnableItem.itemName : string.Empty;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        [field: Tooltip("")]
+        /// <inheritdoc/>
+        [field: Tooltip("Weight value for this specific entry.")]
         [field: Min(0)]
         [field: SerializeField] public int Weight { get; set; } = itemWithRarity.rarity;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        [field: Tooltip("")]
+        /// <inheritdoc/>
+        [field: Tooltip("Weight modifiers to apply whenever this specific entry is used.")]
+        [field: SerializeField] public WeightedModifier[]? WeightedModifiers { get; set; }
+
+        /// <inheritdoc/>
+        [field: Tooltip("Whether this specific entry can be used more than once or not.")]
         [field: SerializeField] public bool SingleUse { get; set; }
 
         /// <summary>
-        ///     TODO.
+        ///     Scrap item corresponding to this specific entry.
         /// </summary>
+        /// <remarks>Deprecated. Should be ignored.</remarks>
         [Space(5.0f)]
         [Header("== DEPRECATED ==")]
-        [Tooltip("(Deprecated) Replace with the desired item's 'itemName' field.")]
+        [Tooltip("(Deprecated) Replace with the desired scrap item's 'itemName' field.")]
+        [Obsolete("Replace with the desired scrap item's 'itemName' field.")]
         public Item? itemToSpawn = itemWithRarity.spawnableItem;
     }
 
     /// <summary>
-    ///     TODO.
+    ///     Represents a scrap item spawner with weighted selection capabilities.
     /// </summary>
     public class ScrapSpawner : NetworkedSpawner<GrabbableObject>, IWeightedScript<ScrapWeightEntry>
     {
         /// <summary>
-        ///     TODO.
+        ///     List of spawned items to be synced with clients.
         /// </summary>
         public NetworkList<ItemInfo> SyncedItems { get; private set; }
 
-        /// <summary>
-        ///     Cached instance of <c>ScrapSpawner</c> as an <c>IWeightedScript</c>, to avoid having to cast.
-        /// </summary>
+        /// <inheritdoc/>
         public IWeightedScript<ScrapWeightEntry> WeightedSelf { get; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        public int[]? CumulativeWeights { get; set; }
+        /// <inheritdoc/>
+        public int[]? CurrentWeights { get; set; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
+        /// <inheritdoc/>
         public int TotalWeight { get; set; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
+        /// <inheritdoc/>
         public bool InitializedWeights { get; set; }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
+        /// <inheritdoc/>
         [field: Space(5.0f)]
         [field: Header("Scrap Spawner")]
-        [field: Tooltip("")]
+        [field: Tooltip("List of weighted entries of type ScrapWeightEntry.")]
         [field: SerializeField] public ScrapWeightEntry[]? WeightedEntries { get; set; }
 
         /// <summary>
-        ///     TODO.
+        ///     Whether to add the current moon's loot pool to the scrap item spawns or not.
         /// </summary>
-        [Tooltip("")]
+        [Tooltip("Whether to add the current moon's loot pool to the scrap item spawns or not.")]
         [FormerlySerializedAs("useMoonScrapSpawns")]
         [SerializeField] private bool addMoonScrapSpawns;
 
         /// <summary>
-        ///     TODO.
+        ///     Minimum value for spawned scrap items.
         /// </summary>
+        /// <remarks>Can be left at <c>-1</c> to disable value overriding.</remarks>
         [Space(5.0f)]
         [Header("Item Properties")]
-        [Tooltip("")]
+        [Tooltip("Minimum value for spawned scrap items. Can be left at '-1' to disable value overriding.")]
         [Min(-1)]
         [SerializeField] private int overrideMinValue = -1;
 
         /// <summary>
-        ///     TODO.
+        ///     Maximum value for spawned scrap items.
         /// </summary>
-        [Tooltip("")]
+        /// <remarks>Can be left at <c>-1</c> to disable value overriding.</remarks>
+        [Tooltip("Maximum value for spawned scrap items. Can be left at '-1' to disable value overriding.")]
         [Min(-1)]
         [SerializeField] private int overrideMaxValue = -1;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether to apply the current scrap value multiplier to spawned scrap items or not.
         /// </summary>
-        [Tooltip("")]
+        [Tooltip("Whether to apply the current scrap value multiplier when overriding spawned scrap item values or not.")]
         [SerializeField] protected bool applyScrapMultiplier = true;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether to allow spawned scrap items to randomly select mesh variants or not.
         /// </summary>
-        [Tooltip("")]
+        [Tooltip("Whether to allow spawned scrap to randomly select mesh variants or not.")]
         [SerializeField] private bool allowMeshVariants = true;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether to allow spawned scrap items to randomly select material variants or not.
         /// </summary>
-        [Tooltip("")]
+        [Tooltip("Whether to allow spawned scrap to randomly select material variants or not.")]
         [SerializeField] private bool allowMaterialVariants = true;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether spawned scrap should fall to the ground or not.
         /// </summary>
         [Space(5.0f)]
         [Header("Position and Rotation")]
-        [Tooltip("")]
+        [Tooltip("Whether spawned scrap should fall to the ground or not.")]
         [SerializeField] private bool fallToGround = true;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether spawned scrap items should have their position slightly offset when falling to the ground upon spawning or not.
         /// </summary>
-        [Tooltip("")]
+        [Tooltip("Whether spawned scrap items should have their position slightly offset when falling to the ground upon spawning or not.")]
         [SerializeField] private bool randomizePosition;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether spawned scrap items should have their defined resting rotation applied or not.
         /// </summary>
-        [Tooltip("")]
+        [Tooltip("Whether spawned scrap items should have their defined resting rotation applied or not.")]
         [SerializeField] private bool applyRestingRotation;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether spawned scrap items should have every <c>AudioSource</c> in it stopped or not.
         /// </summary>
         [Space(5.0f)]
         [Header("Other")]
-        [Tooltip("")]
+        [Tooltip("Whether spawned scrap items should have every audio source in it stopped or not.")]
         [SerializeField] private bool muteSpawnedScrap;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether single item days should be respected when spawning scrap items or not.
         /// </summary>
-        [Tooltip("")]
+                [Tooltip("Whether single item days should be respected when spawning scrap items or not.")]
         [SerializeField] private bool respectSingleItemDay;
 
         /// <summary>
-        ///     TODO.
+        ///     Whether to use the current moon's loot pool as the scrap item spawns or not.
         /// </summary>
+        /// <remarks>Deprecated. Should be ignored.</remarks>
         [Space(10.0f)]
         [Header("== DEPRECATED ==")]
         [Tooltip("(Deprecated) Use 'addMoonScrapSpawns' field instead.")]
+        [Obsolete("Use 'addMoonScrapSpawns' field instead.")]
         [SerializeField] private bool useMoonScrapSpawns;
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc/>
         public override NetworkObject? GetPrefabToSpawn()
         {
+            // Use single item day scrap item spawns instead of weighted selection, if set to respect it.
             if (respectSingleItemDay && TryGetNetworkObject(out NetworkObject itemNetworkObject, SimulateAnomaly.SingleItem))
             {
                 return itemNetworkObject;
             }
 
-            if (WeightedEntries?.Length > 0 && WeightedSelf.TryObtainRandomEntry(out ScrapWeightEntry entry, isSeededRandom
+            if (WeightedEntries?.Length > 0 && WeightedSelf.TryObtainRandomEntry(out ScrapWeightEntry entry, out int _, isSeededRandom
                 ? SeededSelf.GetSeededRandom() : null))
             {
                 if (SearchContent.TryFindItem(out Item item, entry.itemName) && TryGetNetworkObject(out itemNetworkObject, item))
                 {
                     return itemNetworkObject;
                 }
+#pragma warning disable CS0618 // Type or member is obsolete.
                 else if (entry.itemToSpawn != null && (TryGetNetworkObject(out itemNetworkObject, entry.itemToSpawn) || (SearchContent.TryFindItem(out item,
                     entry.itemToSpawn.name, checkObjectName: true) && TryGetNetworkObject(out itemNetworkObject, item))))
+#pragma warning restore CS0618 // Type or member is obsolete.
                 {
                     return itemNetworkObject;
                 }
@@ -198,11 +193,11 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Attempt to obtain the <c>NetworkObject</c> to spawn from a given <c>Item</c>.
         /// </summary>
-        /// <param name="itemNetworkObject"></param>
-        /// <param name="item"></param>
-        /// <returns></returns>
+        /// <param name="itemNetworkObject"><c>NetworkObject</c> of the scrap item to spawn, as an out parameter.</param>
+        /// <param name="item">Item to obtain the <c>NetworkObject</c> off of.</param>
+        /// <returns>Whether a <c>NetworkObject</c> was successfully obtained or not.</returns>
         private static bool TryGetNetworkObject(out NetworkObject itemNetworkObject, Item? item)
         {
             itemNetworkObject = null!;
@@ -210,20 +205,13 @@ namespace itolib.Behaviours.Props
             return item != null && item.spawnPrefab != null && item.spawnPrefab.TryGetComponent(out itemNetworkObject);
         }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc/>
         protected override Transform? GetParentOverride()
         {
             return RoundManager.Instance != null ? RoundManager.Instance.spawnedScrapContainer : null;
         }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
-        /// <param name="item"></param>
-        /// <param name="spawnLocation"></param>
+        /// <inheritdoc/>
         protected override void SpawnPerformed(GrabbableObject? item, TransformInfo spawnLocation)
         {
             if (item == null || !item.IsSpawned || item.itemProperties == null)
@@ -278,9 +266,7 @@ namespace itolib.Behaviours.Props
             }
         }
 
-        /// <summary>
-        ///     TODO.
-        /// </summary>
+        /// <inheritdoc/>
         public override void ClearDestroyed()
         {
             for (int i = PrefabInstances.Count - 1; i >= 0; i--)
@@ -306,7 +292,7 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Initialize weights for every defined <c>ScrapWeightEntry</c> and add the current moon's loot pool, if set to do so.
         /// </summary>
         protected override void Awake()
         {
@@ -317,24 +303,27 @@ namespace itolib.Behaviours.Props
                 WeightedSelf.InitializeWeights();
             }
 
+#pragma warning disable CS0618 // Type or member is obsolete.
             if (useMoonScrapSpawns || addMoonScrapSpawns)
+#pragma warning restore CS0618 // Type or member is obsolete.
             {
-                List<SpawnableItemWithRarity> spawnableScrap = LevelManager.CurrentExtendedLevel.SelectableLevel.spawnableScrap;
-                ScrapWeightEntry[] scrapEntries = new ScrapWeightEntry[spawnableScrap.Count];
-
-                for (int i = 0; i < spawnableScrap.Count; i++)
+                if (StartOfRound.Instance != null && StartOfRound.Instance.currentLevel != null)
                 {
-                    scrapEntries[i] = new(spawnableScrap[i]);
-                }
+                    List<SpawnableItemWithRarity> spawnableScrap = StartOfRound.Instance.currentLevel.spawnableScrap;
 
-                WeightedSelf.AddWeights(scrapEntries);
+                    if (spawnableScrap?.Count > 0)
+                    {
+                        ScrapWeightEntry[] scrapEntries = [.. spawnableScrap.ConvertAll(static itemWithRarity => new ScrapWeightEntry(itemWithRarity))];
+                        WeightedSelf.AddWeights(scrapEntries);
+                    }
+                }
             }
 
             base.Awake();
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Set relevant item values to be synced when spawned.
         /// </summary>
         public override void OnNetworkSpawn()
         {
@@ -351,9 +340,9 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Add a single weighted entry of type <c>ScrapWeightEntry</c>.
         /// </summary>
-        /// <param name="entry"></param>
+        /// <param name="entry">Entry of type <c>ScrapWeightEntry</c> to add.</param>
         public void AddWeightEntry(ScrapWeightEntry entry)
         {
             if (!WeightedSelf.InitializedWeights)
@@ -365,9 +354,9 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Add multiple weighted entries of type <c>ScrapWeightEntry></c>.
         /// </summary>
-        /// <param name="entries"></param>
+        /// <param name="entries">Entries of type <c>ScrapWeightEntry</c> to add.</param>
         public void AddWeightEntries(ScrapWeightEntry[] entries)
         {
             if (!WeightedSelf.InitializedWeights)
@@ -379,9 +368,10 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Remove weights for the weighted entry of type <c>ScrapWeightEntry</c> at the specified index.
         /// </summary>
-        /// <param name="index"></param>
+        /// <remarks>Sets weights to <c>0</c> instead of actually removing them.</remarks>
+        /// <param name="index">Index of the entry of type <c>ScrapWeightEntry</c> to remove.</param>
         public void RemoveWeightEntry(int index)
         {
             if (!WeightedSelf.InitializedWeights)
@@ -393,28 +383,28 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Apply synced item values upon being spawned.
         /// </summary>
-        /// <param name="syncedItem"></param>
-        /// <param name="itemInfo"></param>
-        private void SyncItemValues(GrabbableObject syncedItem, ItemInfo itemInfo)
+        /// <param name="item">Item to be synced.</param>
+        /// <param name="itemInfo">Relevant item values to sync.</param>
+        private void SyncItemValues(GrabbableObject item, ItemInfo itemInfo)
         {
-            _ = StartCoroutine(WaitForItemFall(syncedItem, itemInfo.transformInfo));
+            _ = StartCoroutine(WaitForItemFall(item, itemInfo.transformInfo));
 
-            if (syncedItem.itemProperties != null)
+            if (item.itemProperties != null)
             {
-                if (itemInfo.meshVariant != -1 && syncedItem.TryGetComponent(out MeshFilter itemFilter))
+                if (itemInfo.meshVariant != -1 && item.TryGetComponent(out MeshFilter itemFilter))
                 {
-                    itemFilter.sharedMesh = syncedItem.itemProperties.meshVariants[itemInfo.meshVariant];
+                    itemFilter.sharedMesh = item.itemProperties.meshVariants[itemInfo.meshVariant];
                 }
 
-                if (itemInfo.materialVariant != -1 && syncedItem.TryGetComponent(out MeshRenderer itemRenderer))
+                if (itemInfo.materialVariant != -1 && item.TryGetComponent(out MeshRenderer itemRenderer))
                 {
-                    itemRenderer.sharedMaterial = syncedItem.itemProperties.materialVariants[itemInfo.materialVariant];
+                    itemRenderer.sharedMaterial = item.itemProperties.materialVariants[itemInfo.materialVariant];
                 }
             }
 
-            syncedItem.SetScrapValue(itemInfo.scrapValue);
+            item.SetScrapValue(itemInfo.scrapValue);
 
             if (RoundManager.Instance != null)
             {
@@ -423,12 +413,11 @@ namespace itolib.Behaviours.Props
         }
 
         /// <summary>
-        ///     TODO.
+        ///     Coroutine to wait an extra frame for the item to begin falling before overriding position and rotation.
         /// </summary>
-        /// <param name="item"></param>
-        /// <param name="spawnLocation"></param>
-        /// <returns></returns>
-        private IEnumerator WaitForItemFall(GrabbableObject item, TransformInfo spawnLocation)
+        /// <param name="item">Item that has just spawned.</param>
+        /// <param name="spawnLocation">Position and rotation to apply to the item.</param>
+                private IEnumerator WaitForItemFall(GrabbableObject item, TransformInfo spawnLocation)
         {
             yield return Yielders.WaitForEndOfFrame;
             yield return Yielders.WaitForEndOfFrame;